chore: 코드 리뷰 반영
