+++ src/components/additional-info/car-details/CarColorPicker.module.scss
