+++ src/components/additional-info/car-details/color-picker/ColorPicker.stories.tsx
