+++ src/components/additional-info/details-loading/DetailsLoading.module.scss
