+++ src/components/additional-info/details-loading/DetailsLoading.stories.tsx
