+++ src/components/additional-info/details-loading/DetailsLoading.tsx
