+++ src/remote/queries/additional-info/car-details/useCarSegment.tsx
@@ -1,0 +1,10 @@
+import { useQuery } from '@tanstack/react-query';
+
+import { getCarSegment } from '@remote/api/requests/additional-info/additional-info.get.api';
+import { IAdditionalInfo } from '@remote/api/types/additional-info';
+
+function useCarSegment() {
+  return useQuery<IAdditionalInfo[]>({ queryKey: ['car-segment'], queryFn: getCarSegment });
+}
+
+export default useCarSegment;
