+++ src/remote/queries/additional-info/car-details/useCarType.tsx
@@ -1,0 +1,10 @@
+import { useQuery } from '@tanstack/react-query';
+
+import { getCarType } from '@remote/api/requests/additional-info/additional-info.get.api';
+import { IAdditionalInfo } from '@remote/api/types/additional-info';
+
+function useCarType() {
+  return useQuery<IAdditionalInfo[]>({ queryKey: ['car-type'], queryFn: getCarType });
+}
+
+export default useCarType;
